+++ R/cross-val.R
@@ -14,7 +14,7 @@ ll_binomial <- function(object, withheld_y, withheld_mu) {
 }
 
 ll_gamma <- function(object, withheld_y, withheld_mu) {
-  .shape <- 1.0 / (exp(object$model$par[["ln_phi"]])^2)
+  .shape <- exp(object$model$par[["ln_phi"]])
   stats::dgamma(x = withheld_y, shape = .shape, scale = withheld_mu / .shape, log = TRUE)
 }